pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ragtest"
version = "0.1.0"
description = "RAG evaluation suite — benchmark retrieval accuracy, generation quality, and chunking strategies."
readme = "README.md"
license = {text = "MIT"}
authors = [{name = "Donald Havery", email = "dbhavery@gmail.com"}]
requires-python = ">=3.10"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Testing",
]
dependencies = [
    "click>=8.0",
    "pyyaml>=6.0",
    "rich>=13.0",
]

[project.optional-dependencies]
chromadb = ["chromadb>=0.5"]
dev = ["pytest>=8.0"]

[project.scripts]
ragtest = "ragtest.cli:main"

[project.urls]
Homepage = "https://github.com/dbhavery/ragtest"
Repository = "https://github.com/dbhavery/ragtest"
Issues = "https://github.com/dbhavery/ragtest/issues"

[tool.setuptools.packages.find]
include = ["ragtest*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
